Created scoring system
Added playerWins and playerLoses which are set to zero at the beginning
of the game. When a player loses or wins a round the respective varibale
is incremented by 1 and at the end of the for loop the scores are compared
and a result is logged based on the winner.